tipo de dato de citas recientes se cambio a float

+++ src/main/java/com/ubam/dentcare_plus/dto/cliente/ActividadRecienteDTO.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class ActividadRecienteDTO {
     Integer id;
     LocalDate fecha;
-    BigDecimal costo;
+    Float costo;
     String nombre;
     String estatus;
 }
